requirements.txt
urllib3
bleach
genshi
jinja2
peewee==2.8.8
pillow
pytest
pyyaml
tweepy
requests
selenium
feedparser
readability-lxml
https://github.com/edsu/htmldiff/tarball/master#egg=htmldiff-0.2
